+++ programmable_video/android/src/main/kotlin/twilio/flutter/twilio_programmable_video/PluginHandler.kt
@@ -39,6 +39,7 @@ import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import java.nio.ByteBuffer
+import java.util.ArrayList
 import tvi.webrtc.voiceengine.WebRtcAudioUtils
 
 class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
@@ -106,7 +107,11 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
             "LocalAudioTrack#enable" -> localAudioTrackEnable(call, result)
             "LocalDataTrack#sendString" -> localDataTrackSendString(call, result)
             "LocalDataTrack#sendByteBuffer" -> localDataTrackSendByteBuffer(call, result)
+            "LocalVideoTrack#create" -> localVideoTrackCreate(call, result)
             "LocalVideoTrack#enable" -> localVideoTrackEnable(call, result)
+            "LocalVideoTrack#publish" -> localVideoTrackPublish(call, result)
+            "LocalVideoTrack#unpublish" -> localVideoTrackUnpublish(call, result)
+            "LocalVideoTrack#release" -> localVideoTrackRelease(call, result)
             "RemoteAudioTrack#enablePlayback" -> remoteAudioTrackEnable(call, result)
             "RemoteAudioTrack#isPlaybackEnabled" -> isRemoteAudioTrackPlaybackEnabled(call, result)
             "CameraCapturer#switchCamera" -> switchCamera(call, result)
@@ -141,11 +146,43 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
         VideoCapturerHandler.setTorch(call, result)
     }
 
+    private fun localVideoTrackCreate(call: MethodCall, result: MethodChannel.Result) {
+        val name = call.argument<String>("name")
+            ?: return result.error("MISSING_PARAMS", missingParameterMessage("name"), null)
+        val enabled = call.argument<Boolean>("enable")
+            ?: return result.error("MISSING_PARAMS", missingParameterMessage("enabled"), null)
+        val videoCapturer = call.argument<Map<String, Any>>("videoCapturer")
+            ?: return result.error("MISSING_PARAMS", missingParameterMessage("videoCapturer"), null)
+
+        debug("localVideoTrackCreate => called for $name, enable=$enabled, videoCapturer=$videoCapturer")
+
+        if (TwilioProgrammableVideoPlugin.cameraCapturer == null) {
+            VideoCapturerHandler.initializeCapturer(videoCapturer, result)
+        }
+
+        if (TwilioProgrammableVideoPlugin.localVideoTracks[name] == null) {
+            val localVideoTrack = LocalVideoTrack.create(
+                this.applicationContext,
+                enabled,
+                TwilioProgrammableVideoPlugin.cameraCapturer!!,
+                name
+            ) ?: return result.error(
+                "INIT_ERROR",
+                "Unable to create local video track with name $name",
+                null
+            )
+
+            TwilioProgrammableVideoPlugin.localVideoTracks[name]?.release()
+            TwilioProgrammableVideoPlugin.localVideoTracks[name] = localVideoTrack
+        }
+        result.success(null)
+    }
+
     private fun localVideoTrackEnable(call: MethodCall, result: MethodChannel.Result) {
         val localVideoTrackName = call.argument<String>("name")
-            ?: return result.error("MISSING_PARAMS", "The parameter 'name' was not given", null)
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("name"), null)
         val localVideoTrackEnable = call.argument<Boolean>("enable")
-            ?: return result.error("MISSING_PARAMS", "The parameter 'enable' was not given", null)
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("enable"), null)
 
         TwilioProgrammableVideoPlugin.debug("PluginHandler.localVideoTrackEnable => called for $localVideoTrackName, enable=$localVideoTrackEnable")
 
@@ -157,11 +194,68 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
         return result.error("NOT_FOUND", "No LocalVideoTrack found with the name '$localVideoTrackName'", null)
     }
 
+    private fun localVideoTrackPublish(call: MethodCall, result: MethodChannel.Result) {
+        val localVideoTrackName = call.argument<String>("name")
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("name"), null)
+
+        debug("localVideoTrackPublish => called for $localVideoTrackName")
+
+        val localVideoTrack = TwilioProgrammableVideoPlugin.localVideoTracks[localVideoTrackName]
+                ?: return result.error("NOT_FOUND", "No LocalVideoTrack found with the name '$localVideoTrackName'", null)
+
+        getLocalParticipant()?.publishTrack(localVideoTrack)
+
+        TwilioProgrammableVideoPlugin.localVideoTracks -= localVideoTrackName
+
+        return result.success(null)
+    }
+
+    private fun localVideoTrackUnpublish(call: MethodCall, result: MethodChannel.Result) {
+        val localVideoTrackName = call.argument<String>("name")
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("name"), null)
+
+        debug("localVideoTrackUnpublish => called for $localVideoTrackName")
+
+        val localVideoTrack = getLocalParticipant()
+                ?.localVideoTracks
+                ?.firstOrNull { it.trackName == localVideoTrackName }
+                ?.localVideoTrack
+                ?: return result.error("NOT_FOUND", "No LocalVideoTrack found with the name '$localVideoTrackName'", null)
+
+        getLocalParticipant()?.unpublishTrack(localVideoTrack)
+
+        TwilioProgrammableVideoPlugin.localVideoTracks[localVideoTrackName] = localVideoTrack
+
+        return result.success(null)
+    }
+
+    private fun localVideoTrackRelease(call: MethodCall, result: MethodChannel.Result) {
+        val localVideoTrackName = call.argument<String>("name")
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("name"), null)
+
+        debug("localVideoTrackRelease => called for $localVideoTrackName")
+
+        val preview = TwilioProgrammableVideoPlugin.localVideoTracks[localVideoTrackName]
+        if (preview != null) {
+            preview.release()
+            TwilioProgrammableVideoPlugin.localVideoTracks -= localVideoTrackName
+            return result.success(null)
+        }
+
+        return result.error("NOT_FOUND", "No LocalVideoTrack found with the name '$localVideoTrackName'", null)
+    }
+
     private fun localAudioTrackEnable(call: MethodCall, result: MethodChannel.Result) {
         val localAudioTrackName = call.argument<String>("name")
+<<<<<<< HEAD
             ?: return result.error("MISSING_PARAMS", "The parameter 'name' was not given", null)
         val localAudioTrackEnable = call.argument<Boolean>("enable")
             ?: return result.error("MISSING_PARAMS", "The parameter 'enable' was not given", null)
+=======
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("name"), null)
+        val localAudioTrackEnable = call.argument<Boolean>("enable")
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("enable"), null)
+>>>>>>> programmable-video_v1.0.0
 
         TwilioProgrammableVideoPlugin.debug("PluginHandler.localAudioTrackEnable => called for $localAudioTrackName, enable=$localAudioTrackEnable")
 
@@ -175,14 +269,20 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
 
     private fun localDataTrackSendString(call: MethodCall, result: MethodChannel.Result) {
         val localDataTrackName = call.argument<String>("name")
+<<<<<<< HEAD
             ?: return result.error("MISSING_PARAMS", "The parameter 'name' was not given", null)
         val localDataTrackMessage = call.argument<String>("message")
             ?: return result.error("MISSING_PARAMS", "The parameter 'message' was not given", null)
+=======
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("name"), null)
+        val localDataTrackMessage = call.argument<String>("message")
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("message"), null)
+>>>>>>> programmable-video_v1.0.0
 
         TwilioProgrammableVideoPlugin.debug("PluginHandler.localDataTrackSendString => called for $localDataTrackName")
 
         val localDataTrack = TwilioProgrammableVideoPlugin.roomListener.room?.localParticipant?.localDataTracks?.firstOrNull { it.trackName == localDataTrackName }
-            ?: return result.error("NOT_FOUND", "No LocalDataTrack found with the name '$localDataTrackName'", null)
+                ?: return result.error("NOT_FOUND", "No LocalDataTrack found with the name '$localDataTrackName'", null)
 
         localDataTrack.localDataTrack.send(localDataTrackMessage)
         return result.success(null)
@@ -190,14 +290,20 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
 
     private fun localDataTrackSendByteBuffer(call: MethodCall, result: MethodChannel.Result) {
         val localDataTrackName = call.argument<String>("name")
+<<<<<<< HEAD
             ?: return result.error("MISSING_PARAMS", "The parameter 'name' was not given", null)
         val localDataTrackMessage = call.argument<ByteArray>("message")
             ?: return result.error("MISSING_PARAMS", "The parameter 'message' was not given", null)
+=======
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("name"), null)
+        val localDataTrackMessage = call.argument<ByteArray>("message")
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("message"), null)
+>>>>>>> programmable-video_v1.0.0
 
         TwilioProgrammableVideoPlugin.debug("PluginHandler.localDataTrackSendByteBuffer => called for $localDataTrackName")
 
         val localDataTrack = TwilioProgrammableVideoPlugin.roomListener.room?.localParticipant?.localDataTracks?.firstOrNull { it.trackName == localDataTrackName }
-            ?: return result.error("NOT_FOUND", "No LocalDataTrack found with the name '$localDataTrackName'", null)
+                ?: return result.error("NOT_FOUND", "No LocalDataTrack found with the name '$localDataTrackName'", null)
 
         localDataTrack.localDataTrack.send(ByteBuffer.wrap(localDataTrackMessage))
         return result.success(null)
@@ -205,12 +311,19 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
 
     private fun remoteAudioTrackEnable(call: MethodCall, result: MethodChannel.Result) {
         val remoteAudioTrackSid = call.argument<String>("sid")
+<<<<<<< HEAD
             ?: return result.error("MISSING_PARAMS", "The parameter 'sid' was not given", null)
         val enable = call.argument<Boolean>("enable")
             ?: return result.error("MISSING_PARAMS", "The parameter 'enable' was not given", null)
         TwilioProgrammableVideoPlugin.debug("PluginHandler.remoteAudioTrackEnable => sid: $remoteAudioTrackSid enable: $enable")
+=======
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("sid"), null)
+        val enable = call.argument<Boolean>("enable")
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("enable"), null)
+        debug("remoteAudioTrackEnable => sid: $remoteAudioTrackSid enable: $enable")
+>>>>>>> programmable-video_v1.0.0
         val remoteAudioTrack = getRemoteAudioTrack(remoteAudioTrackSid)
-            ?: return result.error("NOT_FOUND", "No RemoteAudioTrack found with sid $remoteAudioTrackSid", null)
+                ?: return result.error("NOT_FOUND", "No RemoteAudioTrack found with sid $remoteAudioTrackSid", null)
 
         remoteAudioTrack.remoteAudioTrack?.enablePlayback(enable)
         return result.success(null)
@@ -218,17 +331,22 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
 
     private fun isRemoteAudioTrackPlaybackEnabled(call: MethodCall, result: MethodChannel.Result) {
         val remoteAudioTrackSid = call.argument<String>("sid")
+<<<<<<< HEAD
             ?: return result.error("MISSING_PARAMS", "The parameter 'sid' was not given", null)
         TwilioProgrammableVideoPlugin.debug("PluginHandler.isRemoteAudioTrackPlaybackEnabled => sid: $remoteAudioTrackSid")
+=======
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("sid"), null)
+        debug("isRemoteAudioTrackPlaybackEnabled => sid: $remoteAudioTrackSid")
+>>>>>>> programmable-video_v1.0.0
         val remoteAudioTrack = getRemoteAudioTrack(remoteAudioTrackSid)
-            ?: return result.error("NOT_FOUND", "No RemoteAudioTrack found with sid $remoteAudioTrackSid", null)
+                ?: return result.error("NOT_FOUND", "No RemoteAudioTrack found with sid $remoteAudioTrackSid", null)
 
         return result.success(remoteAudioTrack.remoteAudioTrack?.isPlaybackEnabled)
     }
 
     private fun getRemoteAudioTrack(sid: String): RemoteAudioTrackPublication? {
         val remoteParticipants = TwilioProgrammableVideoPlugin.roomListener.room?.remoteParticipants
-            ?: return null
+                ?: return null
 
         var remoteAudioTrack: RemoteAudioTrackPublication?
         for (remoteParticipant in remoteParticipants) {
@@ -238,12 +356,123 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
         return null
     }
 
+<<<<<<< HEAD
     private fun setSpeakerphoneOn(call: MethodCall, result: MethodChannel.Result) {
         val on = call.argument<Boolean>("on")
             ?: return result.error("MISSING_PARAMS", "The parameter 'on' was not given", null)
 
         audioManager.isSpeakerphoneOn = on
         return result.success(on)
+=======
+    private fun setAudioSettings(call: MethodCall, result: MethodChannel.Result) {
+        val speakerphoneEnabled = call.argument<Boolean>("speakerphoneEnabled")
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("speakerphoneEnabled"), null)
+        val bluetoothPreferred = call.argument<Boolean>("bluetoothPreferred")
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("bluetoothPreferred"), null)
+
+        audioSettings.speakerEnabled = speakerphoneEnabled
+        audioSettings.bluetoothPreferred = bluetoothPreferred
+
+        TwilioProgrammableVideoPlugin.audioNotificationListener.listenForRouteChanges(applicationContext)
+
+        applyAudioSettings()
+
+        result.success(null)
+    }
+
+    private fun getAudioSettings(call: MethodCall, result: MethodChannel.Result) {
+        val audioSettingsMap =
+                mapOf(
+                        "speakerphoneEnabled" to audioSettings.speakerEnabled,
+                        "bluetoothPreferred" to audioSettings.bluetoothPreferred
+                )
+        result.success(audioSettingsMap)
+    }
+
+    private fun disableAudioSettings(call: MethodCall, result: MethodChannel.Result) {
+        TwilioProgrammableVideoPlugin.audioNotificationListener.stopListeningForRouteChanges(applicationContext)
+        audioSettings.reset()
+        result.success(null)
+    }
+
+    internal fun applyAudioSettings() {
+        debug("applyAudioSettings")
+        setSpeakerPhoneOnInternal()
+
+        if (!audioSettings.speakerEnabled) {
+            applyBluetoothSettings()
+        }
+    }
+
+    // BluetoothSco being enabled functions similarly to holding Audio Focus when it comes
+    // to external apps audio, if that external app would normally be using the connected
+    // bluetooth device. That is, it prevents the external app from continuing or resuming playback.
+    //
+    // Given this, we only want to turn BluetoothSco on when we are actually using the audio system.
+    internal fun applyBluetoothSettings() {
+        val isConnected = TwilioProgrammableVideoPlugin.isConnected()
+        val anyPlaying = TwilioProgrammableVideoPlugin.audioNotificationListener.anyAudioPlayersActive()
+        debug("applyBluetoothSettings BEGIN =>\n" +
+                "\ton: ${audioSettings.bluetoothPreferred}\n" +
+                "\tscoOn: ${audioManager.isBluetoothScoOn}\n" +
+                "\tconnected: $isConnected\n" +
+                "\tanyPlaying: $anyPlaying")
+        if (isConnected || anyPlaying) {
+            Handler(Looper.getMainLooper()).postDelayed({
+                setBluetoothSco(audioSettings.bluetoothPreferred)
+                audioManager.isBluetoothScoOn = audioSettings.bluetoothPreferred
+                debug("applyBluetoothSettings END => on: ${audioSettings.bluetoothPreferred} scoOn: ${audioManager.isBluetoothScoOn}")
+            }, 1000)
+        }
+    }
+
+    internal fun setBluetoothSco(on: Boolean) {
+        if (on) {
+            audioManager.startBluetoothSco()
+            debug("startBluetoothSco => on: $on\n" +
+                    "\tbluetoothPreferred: ${audioSettings.bluetoothPreferred}\n" +
+                    "\tscoOn: ${audioManager.isBluetoothScoOn}")
+        } else {
+            audioManager.stopBluetoothSco()
+            debug("stopBluetoothSco => on: $on\n\tbluetoothPreferred: ${audioSettings.bluetoothPreferred}\n\tscoOn: ${audioManager.isBluetoothScoOn}")
+        }
+    }
+
+    private fun setSpeakerphoneOn(call: MethodCall, result: MethodChannel.Result) {
+        val on = call.argument<Boolean>("on")
+                ?: return result.error("MISSING_PARAMS", missingParameterMessage("on"), null)
+
+        audioSettings.speakerEnabled = on
+        setSpeakerPhoneOnInternal()
+
+        if (!audioSettings.speakerEnabled && audioSettings.bluetoothPreferred) {
+            applyBluetoothSettings()
+        }
+        return result.success(audioSettings.speakerEnabled)
+    }
+
+    private fun setSpeakerPhoneOnInternal() {
+        val bluetoothProfileConnectionState = BluetoothAdapter.getDefaultAdapter().getProfileConnectionState(BluetoothProfile.HEADSET)
+        debug("setSpeakerPhoneOnInternal => on: ${audioSettings.speakerEnabled}\n bluetoothEnable: ${audioSettings.bluetoothPreferred}\n bluetoothScoOn: ${audioManager.isBluetoothScoOn}\n bluetoothProfileConnectionState: $bluetoothProfileConnectionState")
+
+        // Even if already enabled, setting `audioManager.isSpeakerphoneOn` to true
+        // will reroute audio to the speaker. If using a Bluetooth headset, this will cause audio to
+        // momentarily be routed to the device bottom speaker.
+        //
+        // It has been observed when disconnecting a bluetooth headset that sometimes
+        // the bluetoothProfileConnectionState will still be BluetoothProfile.STATE_CONNECTED
+        // resulting in an edge case where audio will be routed via the receiver rather than the
+        // bottom speaker.
+        if (!audioSettings.bluetoothPreferred ||
+                bluetoothProfileConnectionState != BluetoothProfile.STATE_CONNECTED) {
+            applySpeakerPhoneSettings()
+        }
+    }
+
+    internal fun applySpeakerPhoneSettings() {
+        debug("applySpeakerPhoneSettings => enabled: ${audioSettings.speakerEnabled}")
+        audioManager.isSpeakerphoneOn = audioSettings.speakerEnabled
+>>>>>>> programmable-video_v1.0.0
     }
 
     private fun getSpeakerphoneOn(result: MethodChannel.Result) {
@@ -288,7 +517,7 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
             WebRtcAudioUtils.setWebRtcBasedAcousticEchoCanceler(true)
         }
         val optionsObj = call.argument<Map<String, Any>>("connectOptions")
-            ?: return result.error("MISSING_PARAMS", "Missing 'connectOptions' parameter", null)
+                ?: return result.error("MISSING_PARAMS", "Missing 'connectOptions' parameter", null)
 
         setAudioFocus(true)
 
@@ -406,6 +635,7 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
                 for ((videoTrack) in videoTrackOptions) {
                     videoTrack as Map<*, *> // Ensure right type.
                     val videoCapturerMap = videoTrack["videoCapturer"] as Map<*, *>
+                    val name = videoTrack["name"] as? String ?: ""
 
                     if ((videoCapturerMap["type"] as String) == "CameraCapturer") {
                         VideoCapturerHandler.initializeCapturer(videoCapturerMap, result)
@@ -414,10 +644,20 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
                     }
 
                     if (TwilioProgrammableVideoPlugin.cameraCapturer != null) {
-                        videoTracks.add(LocalVideoTrack.create(this.applicationContext, videoTrack["enable"] as Boolean, TwilioProgrammableVideoPlugin.cameraCapturer!!, videoTrack["name"] as String))
+                        if (name in TwilioProgrammableVideoPlugin.localVideoTracks) {
+                            videoTracks.add(TwilioProgrammableVideoPlugin.localVideoTracks[name])
+                            TwilioProgrammableVideoPlugin.localVideoTracks -= name
+                        } else {
+                            videoTracks.add(LocalVideoTrack.create(this.applicationContext, videoTrack["enable"] as Boolean, TwilioProgrammableVideoPlugin.cameraCapturer!!, videoTrack["name"] as String))
+                        }
                     }
                 }
+<<<<<<< HEAD
                 TwilioProgrammableVideoPlugin.debug("PluginHandler.connect => setting videoTracks to '${videoTracks.joinToString(", ")}'")
+=======
+                debug("connect => setting videoTracks to '${videoTracks.joinToString(", ")}'")
+
+>>>>>>> programmable-video_v1.0.0
                 optionsBuilder.videoTracks(videoTracks)
             }
 
@@ -442,8 +682,14 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
 
     private fun debug(call: MethodCall, result: MethodChannel.Result) {
         val enableNative = call.argument<Boolean>("native")
-            ?: return result.error("MISSING_PARAMS", "Missing 'native' parameter", null)
+                ?: return result.error("MISSING_PARAMS", "Missing 'native' parameter", null)
 
+<<<<<<< HEAD
+=======
+        val enableAudio = call.argument<Boolean>("audio")
+                ?: return result.error("MISSING_PARAMS", "Missing 'audio' parameter", null)
+
+>>>>>>> programmable-video_v1.0.0
         TwilioProgrammableVideoPlugin.nativeDebug = enableNative
         result.success(enableNative)
     }
@@ -460,9 +706,9 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
             // Request audio focus
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
                 val playbackAttributes = AudioAttributes.Builder()
-                    .setUsage(AudioAttributes.USAGE_VOICE_COMMUNICATION)
-                    .setContentType(AudioAttributes.CONTENT_TYPE_SPEECH)
-                    .build()
+                        .setUsage(AudioAttributes.USAGE_VOICE_COMMUNICATION)
+                        .setContentType(AudioAttributes.CONTENT_TYPE_SPEECH)
+                        .build()
                 audioFocusRequest = AudioFocusRequest.Builder(AudioManager.AUDIOFOCUS_GAIN_TRANSIENT)
                     .setAudioAttributes(playbackAttributes)
                     .setAcceptsDelayedFocusGain(true)