+++ programmable_video/lib/src/local_video_track.dart
@@ -30,6 +30,15 @@ class LocalVideoTrack extends VideoTrack {
     videoCapturer._dispose();
   }
 
+  /// Create the local video track.
+  Future<void> create() async {
+    try {
+      await ProgrammableVideoPlatform.instance.createVideoTrack(_toModel() as LocalVideoTrackModel);
+    } on PlatformException catch (err) {
+      throw TwilioProgrammableVideo._convertException(err);
+    }
+  }
+
   /// Set the state of the local video track.
   ///
   /// The results of this operation are signaled to other [Participant]s in the same [Room].
@@ -46,7 +55,7 @@ class LocalVideoTrack extends VideoTrack {
     }
   }
 
-  /// Returns a native widget.
+  /// Publish the local video track.
   ///
   /// By default the widget will be mirrored, to change that set [mirror] to false.
   /// If you provide a [key] make sure it is unique among all [VideoTrack]s otherwise Flutter might send the wrong creation params to the native side.
@@ -60,6 +69,38 @@ class LocalVideoTrack extends VideoTrack {
     );
   }
 
+  /// The results of this operation are signaled to other [Participant]s in the same [Room].
+  /// When a video track is disabled, blank frames are sent in place of video frames from a video capturer.
+  ///
+  /// Throws [MissingParameterException] if [enabled] is not provided.
+  /// Throws [NotFoundException] if no track is found by the name provided (probably means you haven't connected).
+  Future<void> publish() async {
+    try {
+      await ProgrammableVideoPlatform.instance.publishVideoTrack(name);
+    } on PlatformException catch (err) {
+      throw TwilioProgrammableVideo._convertException(err);
+    }
+  }
+
+  /// Unpublish the local video track.
+  ///
+  /// The result of this operation stops signaling to other [Participant]s in the same [Room].
+  Future<void> unpublish() async {
+    try {
+      await ProgrammableVideoPlatform.instance.unpublishVideoTrack(name);
+    } on PlatformException catch (err) {
+      throw TwilioProgrammableVideo._convertException(err);
+    }
+  }
+
+  /// Release the local video track.
+  Future<void> release() async {
+    try {
+      await ProgrammableVideoPlatform.instance.releaseVideoTrack(name);
+    } on PlatformException catch (err) {
+      throw TwilioProgrammableVideo._convertException(err);
+    }
+  }
   /// Create [LocalVideoTrackModel] from properties.
   TrackModel _toModel() {
     final cameraCapturer = videoCapturer as CameraCapturer;