+++ programmable_video/test/local_video_track_test.dart
@@ -16,7 +16,7 @@ void main() {
       expect(
         () => LocalVideoTrack(
           null,
-          CameraCapturer(CameraSource('BACK_CAMERA', false, false, false)),
+          CameraCapturer(const CameraSource('BACK_CAMERA', false, false, false)),
         ),
         throwsAssertionError,
       );
@@ -27,7 +27,7 @@ void main() {
     test('should call interface code to enable the track', () async {
       final localVideoTrack = LocalVideoTrack(
         true,
-        CameraCapturer(CameraSource('BACK_CAMERA', false, false, false)),
+        CameraCapturer(const CameraSource('BACK_CAMERA', false, false, false)),
       );
       await localVideoTrack.enable(false);
 
@@ -37,10 +37,10 @@ void main() {
 
   group('.isEnabled()', () {
     test('should return correct value', () async {
-      final constructionBool = true;
+      const constructionBool = true;
       final localVideoTrack = LocalVideoTrack(
         constructionBool,
-        CameraCapturer(CameraSource('BACK_CAMERA', false, false, false)),
+        CameraCapturer(const CameraSource('BACK_CAMERA', false, false, false)),
       );
       expect(localVideoTrack.isEnabled, constructionBool);
       await localVideoTrack.enable(!constructionBool);