+++ programmable_video/test/programmable_video_test.dart
@@ -12,7 +12,7 @@ void main() {
   group('.debug()', () {
     test('should call interface code to enable native debug', () async {
       final mockInterface = MockInterface();
-      final nativeDebugBool = true;
+      const nativeDebugBool = true;
       ProgrammableVideoPlatform.instance = mockInterface;
       await TwilioProgrammableVideo.debug(native: nativeDebugBool, dart: !nativeDebugBool);
 
@@ -22,7 +22,7 @@ void main() {
 
     test('should call interface code to disable native debug', () async {
       final mockInterface = MockInterface();
-      final nativeDebugBool = false;
+      const nativeDebugBool = false;
       ProgrammableVideoPlatform.instance = mockInterface;
       await TwilioProgrammableVideo.debug(native: nativeDebugBool, dart: !nativeDebugBool);
 
@@ -34,6 +34,8 @@ void main() {
   group('.setSpeakerPhoneOn() & .getSpeakerPhoneOn()', () {
     final mockInterface = MockInterface();
     final on = true;
+    const speakerphoneOn = true;
+    const bluetoothOn = true;
     setUpAll(() => ProgrammableVideoPlatform.instance = mockInterface);
 
     test('should call interface code to enable speakerphone', () async {
@@ -75,7 +77,7 @@ void main() {
       var nativeRequestPermissionsIsCalled = false;
       var nativeCheckPermissionStatusIsCalled = false;
 
-      MethodChannel('flutter.baseflow.com/permissions/methods').setMockMethodCallHandler((MethodCall methodCall) async {
+      const MethodChannel('flutter.baseflow.com/permissions/methods').setMockMethodCallHandler((MethodCall methodCall) async {
         switch (methodCall.method) {
           case 'requestPermissions':
             nativeRequestPermissionsIsCalled = true;
@@ -98,7 +100,7 @@ void main() {
       var nativeRequestPermissionsIsCalled = false;
       var nativeCheckPermissionStatusIsCalled = false;
 
-      MethodChannel('flutter.baseflow.com/permissions/methods').setMockMethodCallHandler((MethodCall methodCall) async {
+      const MethodChannel('flutter.baseflow.com/permissions/methods').setMockMethodCallHandler((MethodCall methodCall) async {
         switch (methodCall.method) {
           case 'requestPermissions':
             if (nativeRequestPermissionsIsCalled) {
@@ -128,7 +130,7 @@ void main() {
       var nativeCheckPermissionStatusIsCalled = false;
       var nativeOpenAppSettingsIsCalled = false;
 
-      MethodChannel('flutter.baseflow.com/permissions/methods').setMockMethodCallHandler((MethodCall methodCall) async {
+      const MethodChannel('flutter.baseflow.com/permissions/methods').setMockMethodCallHandler((MethodCall methodCall) async {
         switch (methodCall.method) {
           case 'requestPermissions':
             nativeRequestPermissionsIsCalled = true;
@@ -156,7 +158,7 @@ void main() {
       final mockInterface = MockInterface();
       ProgrammableVideoPlatform.instance = mockInterface;
 
-      MethodChannel('flutter.baseflow.com/permissions/methods').setMockMethodCallHandler((MethodCall methodCall) async {
+      const MethodChannel('flutter.baseflow.com/permissions/methods').setMockMethodCallHandler((MethodCall methodCall) async {
         switch (methodCall.method) {
           case 'requestPermissions':
             return {1: 1, 7: 1}; // {camera: granted, microphone: granted}