+++ programmable_video_web/lib/src/interop/classes/local_video_track.dart
@@ -2,10 +2,9 @@
 library local_video_track;
 
 import 'dart:html';
-
 import 'package:js/js.dart';
-import 'package:twilio_programmable_video_platform_interface/twilio_programmable_video_platform_interface.dart';
 import 'package:twilio_programmable_video_web/src/interop/classes/video_track.dart';
+import 'package:twilio_programmable_video_platform_interface/twilio_programmable_video_platform_interface.dart';
 
 @JS('Twilio.Video.LocalVideoTrack')
 class LocalVideoTrack extends VideoTrack {