+++ programmable_video_web/lib/src/listeners/local_participant_event_listener.dart
@@ -1,8 +1,7 @@
 import 'dart:async';
 
-import 'package:dartlin/dartlin.dart';
 import 'package:js/js.dart';
-import 'package:twilio_programmable_video_platform_interface/twilio_programmable_video_platform_interface.dart';
+import 'package:dartlin/dartlin.dart';
 import 'package:twilio_programmable_video_web/src/interop/classes/js_map.dart';
 import 'package:twilio_programmable_video_web/src/interop/classes/local_audio_track.dart';
 import 'package:twilio_programmable_video_web/src/interop/classes/local_audio_track_publication.dart';
@@ -15,6 +14,7 @@ import 'package:twilio_programmable_video_web/src/interop/classes/local_video_tr
 import 'package:twilio_programmable_video_web/src/interop/classes/twilio_error.dart';
 import 'package:twilio_programmable_video_web/src/interop/network_quality_level.dart';
 import 'package:twilio_programmable_video_web/src/listeners/base_listener.dart';
+import 'package:twilio_programmable_video_platform_interface/twilio_programmable_video_platform_interface.dart';
 
 class LocalParticipantEventListener extends BaseListener {
   final LocalParticipant _localParticipant;