+++ programmable_video_web/lib/src/listeners/remote_participant_event_listener.dart
@@ -3,7 +3,6 @@ import 'dart:html';
 
 import 'package:dartlin/control_flow.dart';
 import 'package:js/js.dart';
-import 'package:twilio_programmable_video_platform_interface/twilio_programmable_video_platform_interface.dart';
 import 'package:twilio_programmable_video_web/src/interop/classes/js_map.dart';
 import 'package:twilio_programmable_video_web/src/interop/classes/remote_audio_track.dart';
 import 'package:twilio_programmable_video_web/src/interop/classes/remote_audio_track_publication.dart';
@@ -18,6 +17,7 @@ import 'package:twilio_programmable_video_web/src/interop/classes/twilio_error.d
 import 'package:twilio_programmable_video_web/src/interop/network_quality_level.dart';
 import 'package:twilio_programmable_video_web/src/listeners/base_listener.dart';
 import 'package:twilio_programmable_video_web/src/programmable_video_web.dart';
+import 'package:twilio_programmable_video_platform_interface/twilio_programmable_video_platform_interface.dart';
 
 class RemoteParticipantEventListener extends BaseListener {
   final RemoteParticipant _remoteParticipant;
@@ -67,8 +67,7 @@ class RemoteParticipantEventListener extends BaseListener {
     });
   }
 
-  void onTrackDisabledAudio(RemoteAudioTrackPublication publication) =>
-      _remoteParticipantController.add(RemoteAudioTrackDisabled(
+  void onTrackDisabledAudio(RemoteAudioTrackPublication publication) => _remoteParticipantController.add(RemoteAudioTrackDisabled(
         _remoteParticipant.toModel(),
         publication.toModel(),
       ));
@@ -135,6 +134,7 @@ class RemoteParticipantEventListener extends BaseListener {
       ),
     );
 
+    //RMC - the next 2 lines are custom below...
     debug('Remote participant >> Adding video track to remote participants video track list');
     _remoteParticipant.videoTracks.toDartMap()[publication.trackSid] = publication;
   }
@@ -179,7 +179,6 @@ class RemoteParticipantEventListener extends BaseListener {
           audioElement.id = track.name;
           document.body?.append(audioElement);
         });
-
         debug('Attached audio element');
         _remoteParticipantController.add(
           RemoteAudioTrackSubscribed(
@@ -225,6 +224,7 @@ class RemoteParticipantEventListener extends BaseListener {
             remoteAudioTrackModel: audioTrack.toModel(),
           ),
         );
+        //RMC - the next two lines are custom
         debug('Remote participant >> Removing microphone track to remote participants audio track list');
         _remoteParticipant.audioTracks.toDartMap().remove(publication.trackSid);
       },
@@ -245,6 +245,7 @@ class RemoteParticipantEventListener extends BaseListener {
             remoteVideoTrackModel: (track as RemoteVideoTrack).toModel(),
           ),
         );
+        //RMC the next two lines are custom
         debug('Remote participant >> Removing video track to remote participants video track list');
         _remoteParticipant.videoTracks.toDartMap().remove(publication.trackSid);
       },