+++ programmable_video/CHANGELOG.md
@@ -1,4 +1,28 @@
+## 1.0.0
+
+- Integrates updates for mobile platforms from version 0.13.0
+
 ## 1.0.0-alpha.2
+
+- Added `web` as supported platform.
+
+## 1.0.0-alpha.1
+
+- Initial pre-release of the web implementation.
+
+## 0.13.0
+
+- **Android**: Fixed compilation issue for Flutter > 2.12
+- **BREAKING**: Updated permission_handler dependency from 8.3.0 to 9.2.0
+- Added a `create` function to the `LocalVideoTrack` class that will trigger initialization at the native layer.
+- Maintain a map of `LocalVideoTracks` at the native layer to avoid initializing a second track with the same id should the end developer then provide this when connecting.
+- Added a `publishTrack` method to `LocalParticipants` to allow for publishing `LocalVideoTracks` as needed.
+
+## 0.12.1
+
+- **Android**: Fixed the Speaker Not Working when Bluetooth is OFF
+
+## 0.12.0
 
 - Added `web` as supported platform.