+++ programmable_video_platform_interface/CHANGELOG.md
@@ -1,4 +1,19 @@
+## 1.0.0
+
+- Integrates updates for audio settings and video preview from version 0.7.1
+
 ## 1.0.0-alpha.1
+
+- Initial pre-release of the web implementation.
+
+## 0.7.1
+
+- Added `createVideoTrack`.
+- Added `publishVideoTrack`.
+- Added `unpublishVideoTrack`.
+- Added `releaseVideoTrack`.
+
+## 0.7.0
 
 - Initial pre-release of the web implementation.